test(dedupe): consolidate derivatives_discovery tests (5 → 3 files)
Cluster ecb76d4271aa: merge similar_names test files for derivatives_discovery

Files consolidated:
- test_derivatives_discovery.py (NEW): eligibility + discovery_data + both (186 lines)
- test_derivatives_discovery_intx.py: kept (193 lines)
- test_derivatives_discovery_us.py: kept (132 lines)

Removed:
- test_derivatives_discovery_both.py
- test_derivatives_discovery_discovery_data.py
- test_derivatives_discovery_eligibility_and_none.py

All 36 tests pass.
