+++ tests/_triage/dedupe_candidates.yaml
@@ -2586,7 +2586,9 @@ clusters:
     expected_file_delta: -4
     priority: low
     rationale: Consolidate 5 test files into single parametrized file
-    status: pending
+    status: done
+    owner: claude
+    pr_url: https://github.com/Solders-Girdles/GPT-Trader/pull/321
     added: '2026-01-20'
   f08c1e19e6ef:
     type: similar_names
